resources/requirements.txt
pandas
pyarrow
matplotlib
numpy
dask
pypandoc_binary